// models/radios.ts
export class Radio {
    id: number;
    _id: string;
    title: string;
    Imagem_app1: string;
    Imagem_app2: string;
    Imagem_app3: string;
    Imagem_app4: string;
    frequency: string;
    cidade: string;
    parent: boolean;
    RadioRef: string;
    stream_address: string;
    stream_address_app: string;
    advertisingActive: boolean;
    advertisingLink: string;
    advertisingBanner: string;
    ordem: number;
}